Added abstract console implementation.
Now there are 3 consoles each with their own set of commands
