+++ src/main/java/Controller/AbstractConsole.java
@@ -1,0 +1,7 @@
+package Controller;
+
+import java.io.IOException;
+
+public interface AbstractConsole {
+    void run(String input) throws IOException;
+}
